clean profile page, made view all, plusitemwidget
